Make parsing of svlook chaged output more robust.
'svnlook changed' prints the paths that where or will be affected by a commit
preceding with inforamtion of the kind of change for this path. The previous
code assumes that the output consists of two columns, separated with whitespace.
The first with operation type and the second with the path.

This assumption was wrong. In case of renaming there would be (at least) three
columns. First the 'A' (added), than one with an '+' (copied history) and than
the path (I must admit that this form of output is only available with the
--copy-info option to svnlook changed). But svnlook has a fixed format for the
output: first 3 characters which desribe the change, than one space, and the
remainder is the path. This commit changes the parsing by just throwing away the
first 4 characters of an svnlook output line and use the remainder as the path.

It also centralize the clenaup handling of temporary files.


Subversion: common:revision:57