Implement MPI communicator definition backend.
MPI communicators are represented as an OTF2 group of type
OTF2_GROUPTYPE_COMMUNICATOR.  The core does store MPI comms this way so
that the ids for all groups come from the same source. For the adapters
it keeps an MPI comm, not a group.

This also changes the way how an MPI comm is defined by the adapters. The
adapter now passes the array of ranks directly to the core not as an bit
vector.


Subversion: revision:613