Change global location id generation.
This switches the MPI rank and the local thread id position in the global
location id. This should result in better compression in the OTF2 backend.
Because the only event records where an location ID is written are the MPI Send
and Recv events. Which only have MPI Ranks as locations. So putting the MPI rank
into the lower bits keeps the higher bits 0, so that they can be removed with
the OTF2 complression.


Subversion: revision:711