Merge two loops in the MPI communicator unifier.
The creating of the MPI groups and the definition of them and the
referencing MPI communicator was done in two loops (in reverse order,
i.e. first the definition of the communicators and than the groups).
But this is also achievable in one loop.

This removes also the unused attributes global_root_rank and local_id
from the global MPI communicator definition.

And the code is reordered and clearly separated into three phases.


Subversion: revision:1582