Dump Score-P configuration into experiment directory.
The configuration will be dumped into the file "scorep.cfg' into the
experiment directory. It's currently in a format that allows it to be
sourced into the current environment. Ie. the names corresponds to the
environment names. For bash that could be done with:

$ set -a # automatically export all set variables
$ source scorep.cfd
$ set +a # disable it again

This partially closes ticket:262.

This changes also some of the config variable names, namely these of the
profiling subsystem. They are now in the "profiling" config name space.

The case of the namespace and variable names doesn't matter any more. But
the derived environment variable names will be in all-caps. The case for
the bitset and set members doesn't matter anymore too. For the set type the
resulting string list will be converted to lower case. The policy is now
that all namespace, variables, and possible bitset and set member names
should be in lower case. But this commit does not change the latter two.

The namespace and variable names have now a restricted character pool and
length. The length is limited to 32 for each (not total). The namespace
name is limited to alpha numerical characters. The variable name can also
use the underscore, but not at the start or end. This closes ticket:67.

The config system has now a dedicated init and fini interface. The
SCOREP_ConfigRegister() does now only that, register the variables to the
subsytem and apply the default value. The new SCOREP_ConfigApplyEnv()
needs to be called when the environment should be evaluated.

This separation makes it necessary to shuffle the initialization code a
little bit.

There are still memory leaks for the config variables for which the
subsystem allocates memory (currently only the strings) these should be
freed from the adapter when it will be derigistered.



Subversion: revision:1598