Adapted include scheme for public includes from otf2 and untility, namely hat the public headers are included like
#include <scorep/public_header.h>

This leads to a PUBLIC_INC_DIR veariable that is identical for all components
which is desirable as I'm going to extract all common stuff from the various
common.am files into one, located then in the common repository.

Note that the header pomp2_lib.h makes an exception, it is still included like

#include <pomp2_lib.h>

as it is under the control of Opari2. Therefore I added another INCDIR to
scorep_config.dat.


Subversion: revision:1600