Created a libscorep_sion_<paradigm>.la convenience library for the paradigms serial, omp, mpi and mpi_omp and linked them with the corresponding scorep lib. Note that omp and mpi_omp are currently not supported by sionconfig. We go with serial here.
Serial does not register callbacks with OTF2, i.e. OTF2 is responsible for
opening and closing SION files if OTF2_SUBSTRATE_SION is passed to
OTF2_Archive_New. In the MPI case we register callbacks so that we call
sion_paropen_mpi() and sion_parclose_mpi() from the measurement system.

Use the environment variable SCOREP_USE_SION_SUBSTRATE to enable/disable the
use of SION. The default is ''true'' but has no effect if configured without
SION support.


Subversion: revision:1654