Don't link the adapters via the measurement system if they are not needed.
If the adapters are linked, the studio compiler in the shared case will always
execute the function phat_scorep_init() which initializes the measurement
system. This will lead to subsequent errors if the tests initialize and
finalize components on their own. We need to execute phat_scorep_init() in
order to instrument a fake ''main'' function as the studio compiler does not
instrument main for some reason. Note that compiler instrumentation works for
fortran codes only but I don't see a way to prevent execution of
phat_scorep_init() in the c/c++ case.

This will fix the ''make check-serial'' errors in the studio/shared case.


Subversion: revision:1692