Link external libraries to the installed scorep libs, not to the convenience libs.
There was a ''undefined symbols'' error in the shared make check on Juropa
where bfd symbols were not found. libiberty and libz, which are needed by
libbfd, were found though. In contrast to iberty and z bfd is available only
as static library on Juropa. Moving @LIBBFD@ from the compiler adapter's
LIBADD variable to the installed libscorep_*_LIBADD variable solved this
issue.

Thanks to Daniel for reporting this problem.  


Subversion: revision:1835