Define the group with MPI locations as early as possible.
When an MPI program manipulates MPI groups, they are defined in the
measurement system before the unification. But when reading a trace
the list of MPI locations is needed when a MPI group definition is
encountered. But this list is defined at unification time. Which is
too late for MPI groups defined at measurement time. Thus, define
the list of MPI locations as early as possible.

Subversion: revision:2001