Fix mpi unification test after revision:2001.
There was no easy way to define the MPI locations in the test, so I refactored
code in the MPI adapter to re-use the code.

There are also BUG_ON now in the definition system, so that we catch any
problems early. But this change needed that all ranks define the MPI
locations.

Subversion: revision:2005