Don't run the serial and omp metrics testcases with the studio compiler. For this compiler we have no C/C++ compiler adapter, no events are generated, no metrics are recorded, no output is written.
I also disabled the OpenMP testcase. Here we generate output as the POMP2
adapter is active. The test failed though.

Include run_(rusage|papi)_(serial|openmp)_metric_test.sh always in the
distribution. Evene if rusage/papi is not available on the machine generating
the tarball, it might be available on the user's machine.


Subversion: revision:2021