Extend package version scheme with bugfix number and explicit suffix string.
If bugfix is zero, than the third part of the version number is not included
the version string. Ie. it wil still be `1.0-rc1`, not `1.0.0-rc1`.

For the rc1 I would expect something like this in the VERSION file:

package.major=1
package.minor=0
package.bugfix=0
package.suffix=-rc1


Subversion: common:revision:515