Constistently use `@$(am__cd) ... && $(MAKE) $(AM_MAKEFLAGS) ...`
Instead of something like `cd ...; make ...` and remove superfluous '|| exit 1'
at end of recipe lines and remove needless -k handling in recipes, where
we have a fix number of commands to run.

But don't use automake variables in `tools/mpi_wrapgen/Makefile`, which is
not an automake Makefile.


Subversion: revision:2452