New scorep-info tool.
This tool is intended to run on the back-end to provide information
which the user can incorporate into his measurement.

The first info-command is "system-tree" which prints the available system
tree levels of the platform. This will be needed for the upcoming
metric improvements. Where the user can select on which system tree level
he wants to record specific metrics.


Subversion: revision:2963