Add `VERSION` file to configure dependencies.
Please note, because the values are cached from this file, it still
needs a bootstrap to propagate changes to this file into the sources.

But, there is a workaround to disable the caching from the `autoconf` tools.
Create a file name `.autom4te.cfg` next to the top level configures with the
folowing content:

{{{
begin-language: "Autoconf"
args: --no-cache
end-language: "Autoconf"
}}}


Subversion: common:revision:743