Framework to pass backend configure defines to the frontend build.
There is a new `src/config-backend-for-frontend.h` header which contains
all the the `HAVE_BACKEND_` defines via `AC_SCOREP_DEFINE_HAVE`. This header
will be included by `<config.h>` for frontend build. To query the values
in source files which will be build for the backend and the frontend use the
new `HAVE_BACKEND( H )` macro. Which is similar to the known `HAVE( H )` macro
but differs when used in a frontend build. There it does not expands to
`HAVE_##H` but to `HAVE_BACKEND_##H`.


Subversion: revision:3278