Define the debug module name before the `<UTILS_Debug.h>` inclusion.
This also clean-ups the inclusion of `<UTILS_Error.h>` and `<UTILS_Debug.h>`
in headers. For the former, '<SCOREP_ErrorCodes.h>` is the correct one, and the
letter does not make sense in the first place.

It fixes also a wrong `#ifdef __cplusplus` around the `#include <stdbool.h>`
in `SCOREP_Filter.h`.

See common:ticket:6.

Subversion: revision:3464