Fixes ticket:613.
If there is no additional parallelism in a new parallel region, we reuse the TPD object from the previous region to prevent excessive creation of TPD objects in recursive codes. If we reuse a TPD object we must not update it to its parent in SCOREP_Thread_OnThreadJoin.

This is now achieved by maintaining a n_reusages counter in SCOREP_Thread_OnThreadJoin and in SCOREP_Location_GetCurrentCPULocation.

Bert, there are a lot of asserts in this code. Please don't scorep-fy them now as I'm working on the separation of locations and threads. 

Subversion: revision:3478