Allow having no thread parent.
The new TPD implementation has a new feature called `re-use`, which re-uses
the current TPD object in case the new parallel region does not contribute to
additional parallelism (ie. `num_threads == 1`). But in case the first
parallel region already is such a parallel region (ie. `OMP_NUM_THREADS=1`),
than the TPD of the parallel region still does not have a parent (because it
re-uses the TPD from the implicit parallel region). The generic threading layer
does not know about it, but expects that when someone calls
`scorep_thread_get_parent()`, the parallel region does have a parent an bugs on
if not. Which does not hold in the mentioned case. Ignore this bug_on and let the
caller handle NULL pointers if he expects a real parent.

Subversion: revision:4065