Update opari2-related tests as the instrumentation was quite out of date. The hybrid metric test used always the wrong pomp file.
The instrumentation was not only outdated but broken if you used
on-the-fly region registration.

At the moment these tests are useful but the manual update is a
pain. Replacing some of them with installchecks is possible, but is
more time consuming for the developer. Ideas to improve are welcome.


Subversion: revision:4259