Generalize `AC_SCOREP_DEFINE_HAVE` by utilizing `AFS_PACKAGE_BUILD`.
Christian mentioned that it was broken when the macro is called in the
frontend and the backend (e.g. the frontend would also define
`HAVE_BACKEND_FOO`). The build specific macro name is now derived
out of the build name set by `AFS_PACKAGE_INIT` and only if it is a
sub build. I.e. the frontend would define `HAVE_FRONTEND_FOO` and the backend
still defines `HAVE_BACKEND_FOO`.

Subversion: common:revision:1260