AFS_PACKAGE_INIT: remove dependency to `AM_INIT_AUTOMAKE`
The use of the `_AC_SRCDIRS` macro was erroneous. Its usecase is from top-
to sub-configure, not the other way around. Also make the path to the
top-level configure configurable via a new argument and expose it via a new
define named `AFS_PACKAGE_TOP_BUILD`.

The `normalize_path` function was adopted to the case that there is now
always a trailing slash in the SRCDIR.

Subversion: common:revision:1353