Remove usage of absolute build and source paths in tools.
The Score-P instrumenter scorep and the scorep-config tool were using the
absolute path to the build and source directory when operating in the
'--build-check' mode, when they need to access files (e.g., PDT or COBI config
files) or add include flags (e.g., to OPARI2). Convert these paths to be based
on the the calling path of the binary extracted out of 'argv[0]'. As these
paths are only accessed when in '--build-check' mode, we now require that these
tools are called with an path and not via $PATH lookup. Which is also the case
for all callers. Combined with the relative path in `AFS_PACKAGE_SRCDIR`, this
gives also access to the top-level source directory.

It cleans also redundant variable definitions in  `src/tools/config/scorep_config_tool_backend.h.in` up.

See review:72 and review:71.


Subversion: revision:5895