Fix filenames in error and debug messages when they come from an sibling build.
AFS_PACKAGE_SRCDIR is package build dependent. But in case of build-backend
and build-mpi we are building the utils only once for build-backend. That
broke the file name normalization for files which are build in build-mpi.

Pass the AFS_PACKAGE_SRCDIR down as an argument, so that it has the same
source as the __FILE__ itself.

Subversion: common:revision:1790