Remove sanity check whether we are initialized from within an parallel region.
If we are getting this far into an parallel region, we are almost already doomed
anyway. But on the other hand we may call `omp_in_parallel` before the OpenMP
RT is initialized at all, say from an global constructor. As the uninstrumented
application works we should avoid the call to `omp_in_parallel`.

The latter problem seems to happen on 64bit PGI compilers on C++ applications
with an shared Score-P.

Subversion: revision:7120