Update AC_SCOREP_PROG_CC_C99 and status.m4 to autoconf 2.69.
AC_SCOREP_PROG_CC_C99 gets back its `-std=gnu99`. This is required for
MinGW. The reason for not using gnu99 in the first place was
portability. Now, with the bitten clients and the incubation step,
misuse of non-portable GNU extensions should be detected easily.

See review:153.

Subversion: common:revision:1925