Enhanced paradigm handling and properties.
This implements a self-describing and extensible property management for
parallel paradigms in Score-P.

Back in the old days there was only MPI and OpenMP and everybody know about
them. Though the rapid pace of new paradigms made it for the consumer tool hard
to follow. So instead of relying on the knowledge of the consumer, Score-P will
now transport more information about the (possibly) used *parallel* paradigm in
the measurement result.

To achieve, this each adapter will now register the paradigm it is intended to
generate events for with the new 'SCOREP_Paradigms_RegisterParallelParadigm()'
call, which also states the paradigm class of the paradigm and a name.
Additionally properties can be set, if appropriate. For convenience these
properties are split into boolean typed properties and others. Where the former
is expressed via the 'SCOREP_PropertyFlag' parameter to
'SCOREP_Paradigms_RegisterParallelParadigm', the latter via the
'SCOREP_Paradigms_SetStringProperty()' call. Though, as the name implies, only
string typed properties are supported for now.

The class of an paradigm can now be conveniently queried wit the macro
`SCOREP_PARADIGM_TEST_CLASS()`, which is used in the generic threading backends.

Closes ticket:924. Submits review:163.

Subversion: revision:8054