Fix config vars issues for bitsets when 0 is in the set of valid entries
We always write the 'none' entry in bitset config vars into the
'scorep.cfg' file. Example

SCOREP_FILTERING_FORMAT='none','default'

The reason is that this entry has value 0 and this will always be added
to the list. Though 'SCOREP_FILTERING_FORMAT=' would have been the way
to tell that we want the empty set. SCOREP_CUDA_ENABLE have the same
issue with 'no'.

We now support 'no' and 'none' always as possible values to indicate the
empty set. Though these values are now prohibited in the given list of
possible entries, and the value 0 is too.

While at it, change the name "DEFAULT" to "default" as we compare
case-insensitive anyway and fix some typos.

Nothing changes for the user.

Submits review:164.

Subversion: revision:8245