Consolidate the building for the symbol table support.
Instead of deciding on the compiler vendor, the compiler vendor not
only signals it instrumentation flags, but also if it needs the symbol
table. Based on this predicate, we add the symbol table code to the compiler
adapater in one place and not for each compiler.

No functional change intended.

Subversion: revision:8373