Fix linking of serial Fortran jacobi, because of stray reference to the OpenMP API.
The unguarded 'use omp_lib' in main.F90 created a dependency to the OpenMP
runtime on recent PrgEnv-cray, even with `-hnoomp` on the command line. Though
the link step has this flag too, but the linker failes to resolve the reference
to OpenMP. Guard the `use omp_lib` fixes this.

Subversion: revision:8745