Revert revision:8937 ''Simplify the task subsystem callbacks''.
The problem with this changeset is that it breaks
/test/adapters/user/C/run_phase_test.sh for PGI.

The PGI compiler instrumentation interface does not provide any
information which region was exited, see ticket:696. Therefore we explicitly
call SCOREP_Task_Enter for filtered regions.

The first SCOREP_Task_Enter call fails because it is triggered in
between scorep_compiler_initialized and
scorep_subsystems_initialize_location. The former sets
'scorep_compiler_initialized' to true which is the reason why
SCOREP_Task_Enter is called.

SCOREP_Task_Enter is triggered from 'POMP2_Init_reg_xlpth9r14ro9rb_1'
during initialization of the OpenMP adapter. We are not interested in
this function at all. The PGI compiler adapter (every adapter) should be
inactive until SCOREP_InitMeasurement is completed. With the
introduction of SCOREP_IS_MEASUREMENT_PHASE macros, a fix for this
shortcoming is already available in TRY_TUD_sampling.

Till this fix makes it into trunk, revert revision:8937.

Subversion: revision:8969