Merge branches/incubation ''GNU compiler adapter improvements'' into trunk.
See review:182.

When using compiler instrumentation with GNU (not the gcc-plugin but
the -finstrument-functions variant), Cray, or Fujitsu compilers, one
can provide a file containing the symbols to be instrumented. This is
useful when obtaining symbols via 'nm' or 'libbfd' is not an option,
e.g, on Blue Gene systems. The symbol file needs to be specified in
the environment variable SCOREP_NM_SYMBOLS. The accepted format is as
in 'nm -l <executable>'.

Fix bug in obtaining executable names longer than 512 characters when
using the GNU compiler adapter (applies also to Cray and Fujitsu
compilers).

Subversion: revision:9181