Make all SCOREP_ConfigVariable variables `const`.
And also rename most of them to end with `_confvars` as most are now named like
the file they reside in. Remove also some 'scorep_' prefix from static variables.
But only if the declartion and usage are in the same file actually. As this is
not the case for most of the SCOREP_ConfigVariable variables, they kept their
`scorep_` prefix.

The metric servies also had variables which had the `scorep_metrics_` prefix,
renamed to just `scorep_metric_`.

Subversion: revision:9506