Add a new CPU location phase called `PAUSE`.
This is used to supress events at paradigm specific transitions, i.p., the fork
and join phases on the master thread, between the fork/begin and end/join pairs
of events should no samples happen. We use the new PAUSE phase to call into
the subsystems so that we can make this possible. The deactivate(EVENTS) could
not be used, as we would need to distinguish between an intermediate deactivation
and the final one before deactivate(MGMT).

Subversion: revision:9580