Reduce the number of symbols in the compiler symbol table. This reduces the number of collisions and should improve the lookup times (not measured).
From the exposed symbols of hello world programs we roughly take 370
into account. This number is reduced to 15-17, depending an the
language.

From bt-mz_W.4 we used 787 symbols. This number is reduced to 182.

The major reduction is due to filtering of Score-P internal symbols by
comparing the symbol's filename with SCOREP_ABS_TOPLEVEL_SRCDIR.

Subversion: revision:9692