Attempt to fix https://silc.zih.tu-dresden.de/trac-silc/build/trunk-shared-linux/125379 caused by pgf90 15.9 based on Dirk's examination (also removing jacobi_omp_f90_CFLAGS and $(OPENMP_CXXFLAGS) from LD_FLAGS that don't belong to a Fortran program):
I think I found the problem, but did not find a fix so far. The
problem is, that older versions of PGI seem to take the folder of a
source file automatically as include directory and 15.x versions of
PGI do not.  The preinstrumented jacobi examples have *.opari.inc
files in the source directory which are not found by the later PGI
compilers, now.  Adding "-I/<path to fortran examples>" on the compile
line helps.

Subversion: revision:9828