Avoid mismatch between init-struct-list and flags/libs if adapter is not supported.
The cuda, opencl, openacc, and memory adapter are all enabled by default,
also if they are not supported at all. This results in adding the
init-struct but not the flags/libs. Introduce a mockup adapter, to
properly model the absence of an adapters. It also does not show up
in the --help output anymore.

Subversion: revision:10919