doc: Create a new MPI section to hold 'Selection of MPI Groups' and 'Recording MPI Communicator Names'
Neither the 'Trace Buffer Rewind' nor the 'Selective Recording' section is the
right section for the 'Selection of MPI Groups' sub section. Thus create a new
section for MPI performance measurements and convert the 'Recording MPI Communicator Names'
section to a sub section of this new section too.


Subversion: revision:11109