Handle modifications of SHMEM function signatures.
In OpenMPI 2.0 some pointers--passed as argumenmts to SHMEM API functions--are annotated with ''const'' or ''volatile''.
Currently, this results in build failures of the Score-P SHMEM adapter.
The m4 checks were enhanced to check whether the SHMEM installation uses these additional specifiers and the signatures of wrapper code in the Score-P SHMEM adapter are adapted accordingly.
This commit fixes ticket:1072.
Nevertheless, a major update of the SHMEM adapter will be developed in TRY_TUD_openshmem_v1.2 to support functions added in newer releases of the OpenSHMEM standard.

Subversion: revision:11209