Reduce `make dist` dependencies.
We generate `--help` output at `make dist` time to be included in the
documentation. The `--help` output of `scorep-backend-info` and
`scorep-score` is static, i.e., we can prevent `scorep-backend-info`
and `scorep-score` (which depends on libcube4) to be build at `make
dist` time. `scorep-info` still needs to be build at `make dist` time
as it provides the list of all configuration variables that could not
factored out easily.

Subversion: revision:11693