Applying AFS_LINKMODE.
Tested on Piz Daint PrgEnv-gnu
--enable-shared/CRAYPE_LINK_TYPE=static. I couldn't test build-libwrap
as there seems to be no llvm-config available. With this changes make
check succeeds. make installcheck ran into 'disk quota exceeded' but
should not be affected. I just added as many $(LINKMODE_FLAGS) as
needed to get a succesful build for the given configuration. SHMEM
e.g., was not part of this configuration.

Default build on K (--enable-static --disable-shared)
succeeded. Shared builds on K are not supported.

Could not test on JUQUEEN due to maintenance.

Subversion: revision:13320