RFC: Remove Attribute definition from substrate plug-in API.
There are no callbacks to access attributes for this definition
available to plug-ins and no events either. Thus this definition is of
no use to plug-ins at the moment. Therefore I propose that this definition
will be reintroduced to the substrate API with the 'substrate unaware
attributes' feature.

Subversion: revision:13491