Let us find 'llvm-config' again
The fix in revision:13578 had the unintended side effect, that `AC_CHECK_PROGS` was
not executed at all, as the variable had already a value. Fix it, by moving
this fallback into the else branch.

While at it, remove the double-quotes around the ''value-if-not-found''
parameter to `AC_CHECK_PROGS`, as this will already be double-quoted by
Autoconf, i.e., the result was this:

{{{
    scorep_have_llvm_config=""no""
}}}

Subversion: revision:13625