Harden memory API usage heuristic in link step.
CMake likes to add .so files onto the link command as absolute paths. Though
the 'nm'-usage checks for the memory API and the 'weak'-SHMEM wrapping, does
also checks if they use the APIs. But with only link-time wrapping, those
symbols in the .so objects can't be wrapped anymore. Thus remove them from
the object list for the 'nm' command.

Subversion: revision:13665