Workaround buggy code generation of Intel compiler Re TLS access.
At least the Intel compilers 2016-2018 produce cuggy code for this function,
at produce a segmentation fault, when accessing the TLS variable before leaving
the function. The workarround avoids inlining of functions, which was
sufficient to convince the Intel compiler to produce correct code.

No other places could be found, where the Intel compiler is buggy.

Knowledge base entry:

If the application crashes with a SIGSEGV near an instruction like:

{{{
   0x00002aaab5e759a9 <+249>:   mov    0x20d530(%rip),%rbx        # 0x2aaab6082ee0
   0x00002aaab5e759b0 <+256>:   mov    %rbx,0x10(%rsp)
   0x00002aaab5e759b5 <+261>:   mov    0x10(%rsp),%rax
=> 0x00002aaab5e759ba <+266>:   mov    %fs:(%rax),%edx
   0x00002aaab5e759bd <+269>:   dec    %edx
   0x00002aaab5e759bf <+271>:   mov    %edx,%fs:(%rbx)
}}}

Indicatives here is the assignment from `%rbx' to `%rax` via two `mov'
instructions to and from a stack position (via `%rsp`) before accessing the
variable via `%fs:(%rax)`. Than there should also be instructions like:

{{{
   0x00002aaab5e7598a <+218>:   mov    0x20d54f(%rip),%rbx        # 0x2aaab6082ee0
   0x00002aaab5e75991 <+225>:   jmp    0x2aaab5e759b5 <scorep_cupti_callbacks_all+261>
}}}

in this function (before or after does not matter). Where the unconditional jump
is to the `mov` (+261) from the stack into the register, thus skipping the `mov`
from register to the stack.

Subversion: revision:13724