summary: wrap long lines and avoid duplicates
Wraps long lines at 128 margin, but can also be changed per `AFS_SUMMARY`
call. Which is exemplified for the list of installed compiler wrappers.

Wrapping is done with the usual line continuation marker (` \\\n`). Also,
if the key/descr is already too long (>32), than the existing wrapping
includes the continuation marker now too:

Before:
```
        POSIX asynchronous I/O support:
                                yes, using -lrt
```

After:

```
        POSIX asynchronous I/O support: \
                                yes, using -lrt
```

Avoids duplicate summaries between backend and MPI/SHMEM. I.e., lines which
are in MPI/SHMEM and the same in backend wont be printed. This is
controlled by a an optional argument to `AFS_SUMMARY_INIT`, which is the
path to the authoritative summary. The summary hierarchy will be retained,
i.e., if a sub-key differs, then the parent key will also be printed, even
if this does not differ.

Example:

Backend:

```
      OTF2 features:
        SIONlib support:        yes
```

MPI backend:

```
      OTF2 features:
        SIONlib support:        no
```