Reimplement frequency outlier detection using integers only to avoid floating-point inaccuracies.
revision:13097, although fixing the floating point exception, caused the
frequency warnings to occur more frequently than before.

The new implementation calculates the average frequency and ensures
the frequencies to be in the range of 'avg_freq+-threshold' where
threshold is defined as avg_freq/10^6. As CPU clock frequencies are in
the order of 10^9 since 2000 and observed variations are in the order
of 10^3, this heuristic seems reasonable.

Subversion: revision:14336