Remove configure option `--with-extra-instrumentation-flags`
This configure option was introduced to fix
https://silc.zih.tu-dresden.de/trac-silc/ticket/203
https://gitlab.version.fz-juelich.de/perftools-svntogit/scorep-trac/blob/master/ticket/203.tracwiki
for GCC -finstrument-function instrumentation. The ticket wasn't fixed
because the the GCC plug-in instrumentation emerged with which the
problem vanished.

If we want to keep this option, it needs to be made language aware as
language specific instrumentation options will emerge for
Fugaku/Post-K. I consider this of no use.