allocator: Ensure that all pages are paged aligned
If we align the whole allocator by the page size (which is a power-of-two)
than all pages are automatically align too. This than ensures, that the
first allocation from a new page is automatically also aligned.

We do not do the usual 'size + alignment - 1' trick, to still be able to
allocate 4GB, else we may already have an overflow. Thus we deliberately
loose one page in case we need to align.