Generating initial filter files with scorep-score
This is an enhancement of scorep-score, providing the option `-g` to
generate filter file templates to be used as a starting point.

The selection of regions to be included is controlled by two parameters:
1. `-b`: a minimum percent value for the max buffer requirements, as
   basic relevancy check for the trace size
2. `-t`: a maximum threshold for the time/visit in us, as measure for
   the impact on the run-time overhead.
Both parameters have a default value, can be changed by the user and
will produce a warning when used without `-g`.

The filter file is commented extensively to highlight its character as
a starting template, describing the parameters used and including the
remaining column values for each region as additional information.