Fix allocating a page manager for non-CPU locations
In 582788e9 (TRY_PS_InternalMemoryTracking) we started to collect
statistics of the used memory. The page managers for trace events needed
special handling because these could only be accessed from inside the OTF2
memory callbacks. To have access to these page managers they were tried to
be assigned to a Score-P location. But that went horribly wrong, because
it was attached to the current CPU location. That broke all non-CPU
locations in case the CPU location from which the chunks were allocated
flushed.

Now we hook up all of the tracing page managers in a global list for the
only purpose to be accessible when collecting statistics.