summary: Colorize results and different summaries between sub-builds
Color results in the summary. "yes" is green, "no" is red, everything
else is blue. In case of duplicate avoidance of sub-builds and the result
differs to the parent, the key is colored red. Summary is always generated
with color codes, these may be removed in `AFS_SUMMARY_SHOW` or by
`scorep-info show-summary` if not printed to a terminal.

Colorization is done when collating `config.summary` files into
`PACKAGE.summary` file, this way, the `config.summary` files are clean of
color codes and thus still can be read by humans when send in debug
reports.