Unify and fix artificial nodes in profiling
This moves the creation of the THREADS artificial region to a
generalized reroot post-processing step previously used for TASKS. This
avoids erroneous visits to an artificial region.
GPU kernels will use this method to appear in a shared root node GPUS.
This aligns the process in dealing with artificial regions.
The artificial nodes have now the paradigms of their children, which
for example distinguishes CUDA and OpenCl kernels.