Fix NULL substrate initialization for PGI20
PGI 20 identifies as GNUC, but the used __builtin functions return NULL.
This results in an abort from Score-P side.
This commit fixes this by excluding the PGI compiler in the define
condition forcing the use of the else branch.
As these are only optional checks, there shouldn't be any adverse
effects.