cuda: map device index to driver device index
The `CUDA_VISIBLE_DEVICES` variable can be used to restrict and reorder
the  visible devices to the current process. Thus device ID 0 is not
necessarily the first GPU device in the node. To better support
multi-device multi-process measurements, the CUDA adapter implicitly
honors the variable and translates the CUDA device index to the NVIDIA
driver device index when generating the device ID in the location name
`CUDA[<device>:<stream>]`.