Change and enforce parameter order of scorep-score
This changes the parameter parsing to require the cube file to be the
last parameter and to ensure that any option other than 'help' does have
an input file.
This is a preparatory step to allow parameters with optional arguments
by checking the two semantically different parameters first and thereby
simplifying the checking of the remaining '-' parameters.