Switch to single control option for filter gen.
Instead of using multiple options to control the filter generation
heuristic, use an optional parameter (`-g [<list>]`) with a list of
parameters of the form `KEY=VALUE[,KEY=VALUE]*`. This reduces the
number of parameters for this feature in score and at the same time
allows the extension with more options.