[Cray] Provide fallback platform and compilers if type could not be determined
If path `/opt/cray` is available, assume we are on a Cray platform but check
for compilers wrappers cc, CC, ftn. If these exist, use them and switch to
cross-compile mode, even if specific Cray type (XT, XC, etc.) could not be
determined. This is the fallback for not yet supported Cray systems. If the
compiler wrappers don't exist, switch to platform `linux`.

Provides AC_DEFINE `HAVE_PLATFORM_CRAYUNKNOWN` and
AM_CONDITIONAL `PLATFORM_CRAYUNKNOWN`.