Provide macro AFS_EXTERNAL_LIB and AFS_CONSISTENT_DIR
AFS_EXTERNAL_LIB is supposed to replace AC_SCOREP_BACKEND_LIB. It
provides AC_ARG_WITH configure options
  --with-lib<name>=(yes|no|[download|]<path>),
  --with-lib<name>-lib=<path>, and
  [--with-lib<name>-include=<path>]
as well as AC_ARG_VAR precious variable
  LIB<NAME>_LIB, and
  [LIB<NAME>_INCLUDE]

It checks for consistency, if include and lib directories exist, are
absolte paths (AFS_CONSISTENT_DIR) and contain a given header and a
lib<name>.* file. By default, it checks if the library is available in
system locations unless the macro is called from a compute-node
configure (AFS_COMPUTENODE defined) and we detected a cross-compile
platform.

Real autoconf checks are delegated to a check marco the user has to
provide. If a <path> was provided, the check macro can make use of
_afs_lib_CPPFLAGS="-I<incpath>" and _afs_lib_LDFLAGS-"-L<libpath>
-R<libpath>". In addition, _afs_lib_LIBS="-l<name>" is provides. All
three are AC_SUBSTed.

The check is ignored if the download option is chosen. In this case a
download macro is executed, e.g. to create a Makefile that downloads,
builds, and install the library at make time. Then,
`AFS_LIB_DOWNLOAD_CMD <url>` can be used to either wget or curl a
tarball. Variables _afs_lib_MAKEFILE and _afs_lib_PREFIX are
AC_SUBSTed.

Besides the AC_SUBST variables mentioned, other kind of output,
AFS_SUMMARY, AC_SUBST, AC_DEFINE, etc., needs to be done in the check
and download macros.

The default help-string of --with-lib<name> can be overwritten.