Fix skipping of CUDA instrumentation on NVML error
This moves the NVML interaction to create the device mapping into a
helper function allowing for clean handling of NVML errors.
As there are no other negative side effects beside not getting the
mapping information, Score-P can continue with a warning that the device
ids are kept as is without correction based on NVML information.