Change CUDA default/enable set to a full set
This changes the default set of the CUDA_ENABLE options to a reasonable
complete set of options. It doesn't contain runtime as it usually is a
almost redundant copy of the driver set and it uses idle instead of
pure_idle so that a profiling measurement will contain values in the
bytes_put/get metrics.