Switch to default yes for CUDA_ENABLE
As we generally subscribe to the strategy of "first measure everything
of interest, then reduce at need" the default for cuda options should be
the default set - in line with similar behavior as for MPI groups or
compiler instrumentation. If less or no information is required the
CUDA_ENABLE option has to be adapted accordingly.