Avoid construction of debug messages if not needed
Constructing a debug message can cause a non-negligible runtime
penalty, in particular when using the stream-based C++ API.  To
address this, move the check whether the requested debug level is
enabled from the output function to the call site.  This eliminates
the construction overhead in the first place if it is not needed.