cube: Construct region paradigm from group only if it is not the file name
This construction was introduced with the user library wrapper, so that
the wrappers could be distinguished. Note that the file name in this case
is a real one from the header source. In case the region group matches
with the file name, the construction does not add anything, thus skip it.

This will only effect HIP, all previous users of the region group feature
are not effected.