Prevent SegFault in scorep_compiler_demangle
In #206 a case was discovered where a function was found but returned a
value of NULL for the function name. This caused a segmentation fault
when trying to demangle the function name.

With this change, those functions will now be interpreted as
SCOREP_FILTERED_REGION, preventing the crash of Score-P.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>