Add SCOREP_AllocMetric_AddrExists
Right now, most adapters are able to use the available AllocMetric
methods and their existing methods to determine what method should be
called. OMPT is a bit different here. The specification allows
runtimes to call the data_op callback multiple times with the same
pointer and the kind = alloc. This causes a warning each time an
allocation on the same pointer is done. Changing HandleAlloc to
AquireAlloc with an additional check for Realloc / Alloc works, but this
throws a warning each time a new pointer is used.

This commit as a method to check if an allocation is already tracked to
prevent warnings during the data_op callback.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>